profiles: Broadcast profile updates
Add broadcast to indicate profile changes that does not
cause a profile to be selected.

Use the new broadcast to update the Profile Tile more
reliably.

Currently the tile update is triggered by a profile
selection. However, the active profile may change without
triggering a selection. This happens, for instance, on
locale changes.

The Profile Tile does handle locale changes, but is subject
to a race condition. If it handles the locale change before
the profile manager does, the name change will not be seen.

Change-Id: I193454f5100f1ae04dd24d50f25e75ec961f4d68

Conflicts:

	services/java/com/android/server/ProfileManagerService.java