Use sticky broadcast to convey torch state (1/2).
It doesn't really make sense to store the (non-persistent) state of the
torch in a (persistent) setting. As the torch is sending a broadcast
anyway, make that broadcast sticky and use it to convey the state.

Also unify the torch constants (actions, extras, activity names) at a
common place.

Change-Id: If845bef27cc79990794aa748dce18cfea3bac1f5

Conflicts:

	packages/SystemUI/src/com/android/systemui/statusbar/powerwidget/FlashlightButton.java
	policy/src/com/android/internal/policy/impl/keyguard/KeyguardViewManager.java