Wifi: Do not enable configured networks on WPS events.
The Android framework enables the configured networks after WPS
SUCCESS, the supplicant shall consider the networks based on the
priority for the WPA2 handshake after the successful WPS handshake.
This might end up in the station connecting to the AP other than
the current one with which the WPS handshake is done. Moreover,
since framework does not disable the networks and the wpa_supplicant
was the entity that disabled the networks before wps handshake, it
could actually be cleaner to re-enable the networks that were
disabled during the WPS operation within wpa_supplicant. This patch
addresses this by not enabling the configured networks on WPS events.

Bug: 10537481
Change-Id: I2fefc85d0e1f0f1fa96a76940f26aea05219c1ed