fb: clean up our and AOSP code for wifi and carrier label animation
we introduced with this AOSP commit

https://github.com/SlimRoms/frameworks_base/commit/288e3ec70e866105ba61f9ebf259cad29786ad9c

the ability to keep wifi and carrier label up to date on every network
change which happens. The big advantage of this change is that all the
calls for updating the label which are present in PhoneStatusBar are not
needed anymore. We can reduce it to call only when we have animations in progress
or if notifications update and we must decide if the label is still visible or not.
This are way less animation calls then before and on clean AOSP

As well clean up some of our old code which workaround all this calls.

This commit fixes two present minor issues
- carrier label was visible behind notifications (AOSP behaviour but conflicts with our transparency settings)
- due of new updatehandling it happened sometimes that on switching to QS the label was a bit to long visible
- animations are now completly always in sync

PS:
- animate from qs to notification drawer a bit later

PS:
- simplify it more
- fix direct pulldown QS shows sometimes carrier label

Change-Id: I74b80b8b72817da460703635a027f57bb4858f32

Conflicts:

	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBarView.java